requirements.txt
requests==2.31.0
playwright==1.39.0
moviepy==1.0.3
python-dotenv==1.0.1
